pyproject.toml
[tool.poetry]
name = "shoestore"
version = "0.1.0"
description = "WEROFISDF"
authors = ["Bryan Fernandez <bryfern710@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
